split wave_types into modular wave components
